Upgrade @pythnetwork/staking-sdk
The new version of `@pythnetwork/staking-sdk` uses ESM, not CJS.  Because some
modules that are imported are json modules, we are required to use [import
attributes](https://nodejs.org/api/esm.html#import-attributes), or nodejs would
be unable to load the modules.

However, the Realms app has very out-of-date dependencies.  While Realms uses
node v20 which does require import attributes for ESM imports of json files, it
uses a very old and unsupported version of nextjs which includes a version of
webpack & babel which do not support parsing import attributes.

As a stopgap fix, this PR includes a patch which removes the import attributes
from the `@pythnetwork/staking-sdk` files, and configures next to transpile the
package so that the SSR can run in node without them.

Note this is likely going to be a much bigger issue moving forward -- due to
this configuration, this package is unable to load any modern ESM for packages
that must support Node and also import json.  The correct fix would be to
upgrade nextjs, but that is a much larger change that would require someone who
knows this codebase much better than I do.